requirements.txt
discord.py[voice]
aiohttp
PyNaCl